lightcurve api re estructured
