+++ lightcurve/src/lightcurve_api/models/detection.py
