+++ lightcurve/src/lightcurve_api/models/feature.py
