+++ lightcurve/src/lightcurve_api/models/forcephotometry.py
