+++ lightcurve/src/lightcurve_api/models/nondetection.py
