+++ lightcurve/src/lightcurve_api/routes/period.py
@@ -1,8 +1,8 @@
 import pandas as pd
 from fastapi import APIRouter
 
-from core.compute_periodogram import PeriodogramComputer
-from core.harmonics import compute_chi_squared
+from ..services.compute_periodogram import PeriodogramComputer
+from ..services.harmonics import compute_chi_squared
 from ..models.lightcurve_model import LightcurveModel, LightcurveWithPeriod
