+++ lightcurve/src/lightcurve_api/services/compute_periodogram.py
