+++ lightcurve/src/lightcurve_api/services/harmonics.py
